move buffer_view to llama-impl.h
